Changed functionality of the getPriceRange and all methods that use it.
Previously the price range was the maximum price - the minimum price. To
more accurately reflect how a user would expect it to be used or how
they would want it to be used, functionality has been changed so that a
restaurant is filtered out if it has no items within their given price
range.